add cypress and testcafe tutorials

+++ packages/eyes-selenium/tutorial/run.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-
-rm -r ./package
+set -e
+[ -d "./package" ] && rm -r ./package
 mkdir package
 cd ..
 yarn pack
